refactor: Update file structure
